+++ client/src/main/kotlin/com/theovier/athena/client/Launcher.kt
@@ -2,7 +2,7 @@ package com.theovier.athena.client
 
 import com.badlogic.gdx.backends.lwjgl3.Lwjgl3Application
 import com.badlogic.gdx.backends.lwjgl3.Lwjgl3ApplicationConfiguration
-import com.theovier.athena.client.ecs.prefabs.loaders.SpriteComponentLoader
+import com.theovier.athena.client.ecs.prefabs.loaders.components.SpriteComponentLoader
 import ktx.assets.async.AssetStorage
 import org.koin.core.context.GlobalContext.startKoin
 import org.koin.dsl.module
