+++ client/src/main/kotlin/com/theovier/athena/client/ecs/prefabs/loaders/components/MovementComponentLoader.kt
@@ -1,4 +1,4 @@
-package com.theovier.athena.client.ecs.prefabs.loaders
+package com.theovier.athena.client.ecs.prefabs.loaders.components
 
 import com.badlogic.ashley.core.Component
 import com.badlogic.gdx.utils.JsonValue
