--- client/src/main/kotlin/com/theovier/athena/client/ecs/prefabs/wrappers/SerializableParticleEffect.kt
@@ -1,8 +1,0 @@
-package com.theovier.athena.client.ecs.prefabs.wrappers
-
-import com.badlogic.gdx.graphics.g2d.ParticleEffect
-
-/*
-    have to use this as workaround for https://github.com/theovier/athena/issues/8
- */
-class SerializableParticleEffect : ParticleEffect()
